Pull all compiled event values into compiled record
This allows us to drop all of the values prefetched from
run-task-lifecycle and make the pipeline functions more consistent in
terms of arguments taken (i.e. just compiled, event).

Then we can genericise some of the lifecycle-invoke functions as they
all take the same args. This will eventually allow us to get rid of the
list destructuring and later apply in
lifecycle-invoke/restartable-invokation

I then switched all of the intial pipeline setup functions to take the
event as the last argument, as that’s what the task-lifecycle does.
This makes switching around any lifecycle functions a breeze at any
stage.

Lastly, I removed merge from all of the invokes, most of them already
merged in their lifecycle function, and this was leading to duplicate
merges.