Improvements to how orphaned-peers are managed
Key orphaned-peers by group-id for simplicity.
We then allow add-virtual-peers to automatically join the
orphaned-peers, preventing the need for a backoff mechanism. In order
so that we don’t end up with orphaned peers being maintained, we also
maintain the groups that have left under :left. This allows us to
ignore add-virtual-peer for groups that have already left.