Don't throw an error on JVM shutdown hook if Aeron directory is gone.
After turning up the log levels sufficiently high, we've revealed
that deleting the Aeron directories on shutdown virtually always
throws a not-found exception. We've never seen this before since
JVM shutdown hook threads suppress their exceptions, and we didn't
have the log levels up high enough to see it.

This patch suppresses the exception -- which looks scary when
seen from the console, but is ultimately harmless.